fix permission issues while importing - add "reload library" functionality

+++ src/worker/index.ts
@@ -1,10 +1,13 @@
-import { forceAddDirectory, removeDirectory, tryAddDirectory } from './library';
+import { forceAddDirectory, removeDirectory, tryAddDirectory, updateFiles } from './library';
 import { getStore } from './store';
 import { onMessage } from './utils';
 
 onMessage(async ({ data }) => {
   if (data.message === 'getStore') {
     return getStore();
+  }
+  if (data.message == 'reloadLibrary') {
+    return updateFiles();
   }
   if (data.message === 'removeLibraryDirectory') {
     return removeDirectory(data.id);
